remote_sensing_api-main/requirements.txt
aenum==3.1.15
affine==2.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dataclasses-json==0.6.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
gunicorn==22.0.0
geopandas==1.0.0
idna==3.7
marshmallow==3.21.3
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
rasterio==1.3.10
requests==2.32.3
requests-oauthlib==2.0.0
sentinelhub==3.10.2
shapely==2.0.4
six==1.16.0
snuggs==1.4.7
sqlparse==0.5.0
tifffile==2024.6.18
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utm==0.7.0
scikit-learn==1.5.2
prophet==1.1.6